Refactor CloudShellCredential
